# openapi-spec.yml
openapi: 3.0.1
info:
  title: sample golang application
  description: demo application
  version: v1
servers:
- url: http://localhost:8080/v1
  description: Main production server
paths:
  /demo/service/v1/sprints/:sprintId:
    get:
      tags:
      - sprints
      summary: Returns a Sprint
      description: Returns a Sprint
      operationId: GetSprint
      parameters:
      - name: sprintId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sprint'
        default:
          description: Default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
      - sprints
      summary: Updates a Sprint
      description: Updates a Sprint
      operationId: UpdateSprint
      parameters:
      - name: sprintId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sprint'
      responses:
        200:
          description: Default response
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
      - sprints
      summary: Delete Sprint
      description: Delete Sprint
      operationId: DeleteSprint
      parameters:
      - name: sprintId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: Success response
  /demo/service/v1/sprints:
    post:
      tags:
      - sprints
      summary: Creates Sprint
      description: Creates Sprint
      operationId: CreateSprint
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sprint'
      responses:
        200:
          description: Default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sprint'
        default:
          description: Default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /demo/service/v1/sprints/:sprintId/stories/:storyId:
    get:
      tags:
      - sprints
      summary: Returns a Sprint
      description: Returns a Sprint
      operationId: GetSprint
      parameters:
      - name: sprintId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sprint'
        default:
          description: Default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
      - sprints
      summary: Updates a Sprint
      description: Updates a Sprint
      operationId: UpdateSprint
      parameters:
      - name: sprintId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sprint'
      responses:
        200:
          description: Default response
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
      - sprints
      summary: Delete Sprint
      description: Delete Sprint
      operationId: DeleteSprint
      parameters:
      - name: sprintId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: Success response
  /demo/service/v1/sprints/:sprintId/stories:
    post:
      tags:
      - sprints
      summary: Creates Sprint
      description: Creates Sprint
      operationId: CreateSprint
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sprint'
      responses:
        200:
          description: Default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sprint'
        default:
          description: Default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...